e-hal: shm: Defer checking shm table integrity to after e_init()
Defer shm table integrity check until it is actually used.
Previously e_shm_init() would fail if the magic field did not match.
This means e_init() would also fail so the API did not provide a way
to clear the shm table.

Instead defer integrity check to when the table is accessed.
Also, make it so that the table is cleared when e_reset_system() is called.