e-hal: Import epiphany.h from kernel sources
We need *one* header file for building e-hal. By importing it into the
e-hal sources  we get rid of the need to clone the kernel sources when
building the toolchain using the epiphany-sdk build scripts.

The only downside is that the header files need to stay in sync.

src/e-hal/
	* src/epiphany-shm-manager.c: Use in-tree epiphany.h
	rather than kernel version.
	* src/epiphany.h: New file.

Signed-off-by: Ola Jeppsson <ola@adapteva.com>