Fix handling of unknown breakpoint packet kinds
The insert/remove breakpoints code is assuming that future Zn packets
follow the "z%1d,%x,%1ud" format, which is not a good assumption.

Also, if any other zn/Zn packet is created in the future, given 'n' is
not covered in the MpType enum, we'll end up returning "E01", but we
should instead return "".

So this commit makes us only recognize z0/Z0 packets, and bails out
earlier for any other zn/Zn packet.

2016-06-22  Pedro Alves  <pedro@palves.net>

	* src/GdbServer.cpp (GdbServer::rspRemoveMatchpoint)
	(GdbServer::rspInsertMatchpoint): Return unknown packet for any
	matchpoint other than z0/Z0.

Signed-off-by: Pedro Alves <pedro@palves.net>