Replace the pendingStops set with a couple flags in the Thread object
The main idea here is make the "wait for stops" part of
GdbServer::rspVCont be independent of the last parsed "threadActions"
object.

 - We'll want to factor out that "wait or stops" code for use in other
   cases.

 - In non-stop mode, vCont and the "wait for stops" are decoupled.  We
   can see multiple vCont packets before any thread stops, even.

And for that, we need to keep track of what was the vCont last action
applied to each thread.

Since we're adding new flag to Thread, we might as well convert the
pendingStops set to a flag as well.  Since we'll normally have to
iterate the whole list of threads to check whether any halted, keeping
a separate list is questionable.  In any case, the current container
is a "std::set" (red black tree), sorted by thread id, which I very
much doubt is the best algorithm for this..  Thus, if we ever consider
optimizating the list of already-stopped-threads, I believe we'd do
something different anyway.

2016-06-22  Pedro Alves  <pedro@palves.net>

	* src/GdbServer.cpp (GdbServer::rspAttach): Call markAllStopped.
	(GdbServer::rspReportException): Call markPendingStops.
	(GdbServer::rspSuspend): Remove parameter.  Instead fetch the
	thread's last action from the thread object directly.
	(GdbServer::rspSuspend): Call markAllStopped.
	(GdbServer::rspVCont): Get a thread's last action and pending
	status from the thread object directly.
	(GdbServer::pendingStop): Delete.
	(GdbServer::markPendingStops): Remove ProcessInfo pointer
	parameter.  Always work with the current process.  Get/set a
	thread's last action and pending status from/in the thread object
	directly.
	(GdbServer::removePendingStop): Delete.
	(GdbServer::markAllStopped): New method.
	(GdbServer::doContinue): Mark all threads stopped.
	(GdbServer::resumeAllThreads): Skip continued threads with a
	pending status.
	* src/GdbServer.h (GdbServer) <mPendingStops>: Remove field.
	<markPendingStops, rspSuspend>: Change prototype.
	<removePendingStop>: Remove.
	<markAllStopped>: New.
	* src/Thread.cpp (Thread::Thread): Initialize mPending.
	* src/Thread.h (Thread) <isPending, setPending, clearPending,
	setLastAction, lastAction>: New methods.
	<mLastAction, mPending>: New fields.

Signed-off-by: Pedro Alves <pedro@palves.net>