Fix spurious resumes causing random odd behavior
This patch fixes this internal error:

  Thread 1 hit Breakpoint 1, main () at src/e_prime.c:64
  64            number += (2*16);
  (gdb) inline-frame.c:343: internal-error: void skip_inline_frames(ptid_t): Assertion `find_inline_frame_state (ptid) == NULL' failed.
  A problem internal to GDB has been detected,
  further debugging may prove unreliable.
  Quit this debugging session? (y or n)

and likely other random odd behaviors.

This is usually a consequence of a target/server-side bug: it'll happen
if the target sends back a thread stop event for a thread that gdb did
not resume in the first place.  And this is exactly what happens with
e-server.  The problem is here, in a path that is missing a return
statement:

 In file included from ../e-server/src/GdbServer.cpp:42:0:
 ../e-server/src/GdbServer.h: In member function ‘bool GdbServer::vContTidAction::matches(int, int) const’:
 ../e-server/src/GdbServer.h:188:5: warning: control reaches end of non-void function [-Wreturn-type]
      }
      ^

This function is used by the vCont packet handling, to determine whether a
thread matches a vCont action.  Due to the bug, sometimes the wrong thread
would match and be incorrectly resumed.

e-server/
	* src/GdbServer.h (GdbServer::vContTidAction::matches): Add
	missing return path.