e-hal: Add back "public" ee_set_core_config
This is ugly, but easiest way to support OMPi 2.0.0 atm.
OMPi has its own e_load_group () that still uses SREC files,
(not supported anymore). It uses ee_set_core_config which was
never intended to be part of the public API, (hence the ee_ prefix).
This is what we get for exporting private API functions.

/
	* e-hal/src/e-loader.c (ee_set_core_config): Rename ...
	(_ee_set_core_config): ... to this.
	(ee_set_core_config): "New" function. Added as OMPi 2.0.0
	compatibility hack. DO NOT USE.

Signed-off-by: Ola Jeppsson <ola@adapteva.com>