e-hal: Rudimentary e-server (gdbserver) awareness
If EHAL_GDBSERVER environment variable is set to non-empty, start cores
in halted debugstate, that is: write 1 to DEBUGCMD before sending sync
interrupt in _e_default_start_group().

This will leave the cores halted in _start() until a user attaches with GDB
to e-server. Now it's possible to debug an Epiphany program, started by
e-hal, from the beginning.

/
	* e-hal/src/epiphany-hal.c (gdbserver_attached_p): New
	function.
	(_e_default_start_group): Halt cores, before sending SYNC if
	EHAL_GDBSERVER env var is set.
	(static int e_halt_group): New function.

Signed-off-by: Ola Jeppsson <ola@adapteva.com>