e-lib: Work around API / GCC interrupt attribute differences
Interrupt Service Routine functions must be declared with the interrupt
attribute. This requires a void function with no arguments. But this
clashes with the sighandler_t type, which is intended for signals.

Add a new e_irqhandler_t type that matches the compilers expectations. In
e_irq_attach, change the handler type to e_irqhandler_t.

Add a cover macro for e_irq_attach, that casts to e_irqhandler_t to
maintain API compatibility.

/
	* e-lib/include/e_ic.h (e_irqhandler_t): New typedef.
	(e_irq_attach): Update function prototype. Add cover macro
	that casts handler argument to e_irqhandler_t.
	* e-lib/src/e_irq_attach.c: Change handler type to
	e_irqhandler_t.

Signed-off-by: Ola Jeppsson <ola@adapteva.com>